Added ell to PVSS tables, also changed some usizes to u8s

+++ crates/aptos-crypto/src/utils.rs
@@ -26,7 +26,7 @@ where
 }
 
 /// Asserts that the given value is a power of two.
-pub fn assert_power_of_two(n: usize) {
+pub fn assert_power_of_two(n: u64) {
     assert!(
         n.is_power_of_two(),
         "Parameter must be a power of 2, but got {}",
