+++ crates/aptos-dkg/benches/range_proof.rs
@@ -29,7 +29,7 @@ const BLS12_381: &str = "bls12-381";
 const BATCH_SIZES: [usize; 11] = [1, 3, 7, 15, 31, 63, 127, 255, 511, 1023, 2047];
 
 /// WARNING: These are the relevant bit widths we want benchmarked to compare against Bulletproofs
-const BIT_WIDTHS: [usize; 4] = [8, 16, 32, 64];
+const BIT_WIDTHS: [u8; 4] = [8, 16, 32, 64];
 
 fn bench_groups(c: &mut Criterion) {
     bench_range_proof::<Bn254, UnivariateDeKART<Bn254>>(c, BROKEN_DEKART_RS_SCHEME_NAME, BN254);
@@ -57,7 +57,7 @@ fn bench_range_proof<E: Pairing, B: BatchedRangeProof<E>>(
 
     let l = std::env::var("L")
         .ok()
-        .and_then(|s| s.parse::<usize>().ok());
+        .and_then(|s| s.parse::<u8>().ok());
     let n = std::env::var("N")
         .ok()
         .and_then(|s| s.parse::<usize>().ok());
@@ -80,7 +80,7 @@ fn bench_range_proof<E: Pairing, B: BatchedRangeProof<E>>(
 
 fn bench_verify<E: Pairing, B: BatchedRangeProof<E>>(
     group: &mut BenchmarkGroup<WallTime>,
-    ell: usize,
+    ell: u8,
     n: usize,
 ) {
     group.bench_function(
@@ -106,7 +106,7 @@ fn bench_verify<E: Pairing, B: BatchedRangeProof<E>>(
 
 fn bench_prove<E: Pairing, B: BatchedRangeProof<E>>(
     group: &mut BenchmarkGroup<WallTime>,
-    ell: usize,
+    ell: u8,
     n: usize,
 ) {
     group.bench_function(
